requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.2
django-ckeditor==6.7.0
django-cors-headers==3.14.0
django-crontab==0.7.1
django-js-asset==2.1.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
fdfs-client-py==1.2.6
idna==3.4
kombu==5.3.1
mutagen==1.41.1
Pillow==10.1.0
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.6
